poker_ex v0.0.1 PokerEx.Card

Types

rank()

rank ::
  :two |
  :three |
  :four |
  :five |
  :six |
  :seven |
  :eight |
  :nine |
  :ten |
  :jack |
  :queen |
  :king |
  :ace |
  :joker

suit()

suit :: :spades | :diamonds | :hearts | :clubs

t()

t :: %PokerEx.Card{rank: rank | nil, suit: suit | nil}

Functions

from_atom(atom)

from_atom(atom) :: PokerEx.Card.t

Takes in an atom describing a card and returns a corresponding Card struct

Examples

iex> PokerEx.Card.from_atom(:ace_of_spades)
    %PokerEx.Card{rank: :ace, suit: :spades}

sort_by_rank(hand)

sort_by_rank([PokerEx.Card.t]) :: [PokerEx.Card.t]

Sorts cards based on their rank

Examples

iex> alias PokerEx.Card
    iex> cards = [:ten_of_hearts, :two_of_spades, :ace_of_diamonds, :ten_of_clubs, :three_of_hearts] |> Enum.map(&Card.from_atom/1)
    iex> Card.sort_by_rank(cards)
    [%PokerEx.Card{rank: :ace, suit: :diamonds}, %PokerEx.Card{rank: :ten, suit: :hearts}, %PokerEx.Card{rank: :ten, suit: :clubs},
    %PokerEx.Card{rank: :three, suit: :hearts}, %PokerEx.Card{rank: :two, suit: :spades}]

value(card)

value(PokerEx.Card.t) :: pos_integer

Returns the numerical value of a card for comparison purposes.

value_to_rank(number)

value_to_rank(pos_integer) :: rank

Takes in an integer value from 2 to 14 and returns the corresponding card rank

Examples

iex> PokerEx.Card.value_to_rank(14)
    :ace