poker_ex v0.0.1 PokerEx.Evaluator

Implements business logic for determining a winning hand in Hold ‘Em.

Types

hand()

hand :: PokerEx.Hand.t

Functions

ace_to_five_straight?(hand_pool)

ace_to_five_straight?([PokerEx.Card.t]) :: boolean

check_for_flush(hand)

check_for_flush(hand) :: hand

check_for_n_kind(hand)

check_for_n_kind(hand) :: hand

check_for_straight(hand)

check_for_straight(hand) :: hand

check_type(hand)

check_type(hand) :: hand

evaluate_hand(player_hand, table)

evaluate_hand([PokerEx.Card.t], [PokerEx.Card.t]) :: hand

Chooses the best five-card hand from a set of seven cards built from a player’s hand and the hand dealt on the table.

Examples

iex> alias PokerEx.Card
    iex> alias PokerEx.Evaluator
    iex> player_hand = [:queen_of_diamonds, :queen_of_hearts] |> Enum.map(&Card.from_atom/1)
    iex> table = [:two_of_clubs, :four_of_spades, :ten_of_diamonds, :ace_of_hearts, :three_of_spades] |> Enum.map(&Card.from_atom/1)
    iex> Evaluator.evaluate_hand(player_hand, table)
    %PokerEx.Hand{best_hand: [%PokerEx.Card{rank: :ace, suit: :hearts},
    %PokerEx.Card{rank: :four, suit: :spades},
    %PokerEx.Card{rank: :queen, suit: :diamonds},
    %PokerEx.Card{rank: :queen, suit: :hearts},
    %PokerEx.Card{rank: :ten, suit: :diamonds}],
    hand: [%PokerEx.Card{rank: :queen, suit: :diamonds},
    %PokerEx.Card{rank: :queen, suit: :hearts},
    %PokerEx.Card{rank: :two, suit: :clubs},
    %PokerEx.Card{rank: :four, suit: :spades},
    %PokerEx.Card{rank: :ten, suit: :diamonds},
    %PokerEx.Card{rank: :ace, suit: :hearts},
    %PokerEx.Card{rank: :three, suit: :spades}], hand_type: :one_pair,
    has_flush_with: nil,
    has_n_kind_with: [%PokerEx.Card{rank: :queen, suit: :diamonds},
    %PokerEx.Card{rank: :queen, suit: :hearts}], has_straight_with: nil,
    score: 152, type_string: "a Pair of Queens"}

score(hand)

score(hand) :: hand