poker_ex v0.0.1 PokerEx.Room.Updater

Provides convenience functions for updating specific attributes of the Room state individually. This module can be used from other modules working with the Room state struct to facilitate updating the state in a pipeline.

Types

player()

player :: String.t

Functions

active(room)

active(PokerEx.Room2.t) :: PokerEx.Room2.t

Sets the list of players under the seating attribute to active. Used when setting up a round.

Examples

iex> room = %Room{seating: [{"A", 0}, {"B", 1}]}
iex> Updater.active(room)
%Room{active: [{"A", 0}, {"B", 1}], seating: [{"A", 0}, {"B", 1}]}

active(room, player)

active(PokerEx.Room2.t, player) :: PokerEx.Room2.t

Removes a single player from the active list. Note that active/2 removes a player whereas active/1 sets the active list at game start.

Examples

iex> room = %Room{active: [{"A", 0}, {"B", 1}]}
iex> Updater.active(room, "B")
%Room{active: [{"A", 0}]}

advance_active(room)

advance_active(PokerEx.Room2.t) :: PokerEx.Room2.t

Rotates the active list to keep the head of the list current as the player whose turn it is.

Examples

iex> room = %Room{active: [{"A", 0}, {"B", 1}, {"C", 2}]}
iex> Updater.advance_active(room)
%Room{active: [{"B", 1}, {"C", 2}, {"A", 0}]}

all_in(room, player)

all_in(PokerEx.Room2.t, player) :: PokerEx.Room2.t

Updates the list of players who have gone all in during the current game while removing the specified player from the list of active players

Examples

iex> room = %Room{all_in: ["A"], active: [{"C", 2}, {"D", 3}]}
iex> Updater.all_in(room, "C")
%Room{all_in: ["C", "A"], active: [{"D", 3}]}

blinds(room)

blinds(PokerEx.Room2.t) :: PokerEx.Room2.t

Updates the blinds between rounds and before the first game played at a table. If both the current_big_blind and current_small_blind data attributes are nil, they are set to 1 and 0 by default, respectively. If the big blind is currently equal to the length of the seating list minus 1, the big blind position gets reset to 0. If the big blind is less than the small blind (in terms of seating position), the small blind gets reset to 0 while the big_blind moves up to 1.

Examples

iex> room = %Room{current_big_blind: nil, current_small_blind: nil}
iex> room = Updater.blinds(room)
iex> {room.current_small_blind, room.current_big_blind}
{0, 1}

iex> room = %Room{current_big_blind: 1, current_small_blind: 0, seating: [{"A", 0}, {"B", 1}, {"C", 2}, {"D", 3}]}
iex> room = Updater.blinds(room)
iex> {room.current_small_blind, room.current_big_blind}
{1, 2}

iex> room = %Room{current_big_blind: 3, current_small_blind: 2, seating: [{"A", 0}, {"B", 1}, {"C", 2}, {"D", 3}]}
iex> room = Updater.blinds(room)
iex> {room.current_small_blind, room.current_big_blind}
{3, 0}

iex> room = %Room{current_big_blind: 0, current_small_blind: 3, seating: [{"A", 0}, {"B", 1}, {"C", 2}, {"D", 3}]}
iex> room = Updater.blinds(room)
iex> {room.current_small_blind, room.current_big_blind}
{0, 1}

iex> room = %Room{current_big_blind: 2, current_small_blind: 1, seating: [{"A", 0}, {"B", 1}, {"C", 2}, {"D", 3}]}
iex> room = Updater.blinds(room)
iex> {room.current_small_blind, room.current_big_blind}
{2, 3}

call_amount(room, amount)

call_amount(PokerEx.Room2.t, pos_integer) :: PokerEx.Room2.t

Updates the bet required to call or raise. Does nothing if the amount specified is less than or equal to the current amount to be called in the to_call attribute

Examples

iex> room = %Room{to_call: 20}
iex> Updater.call_amount(room, 40)
%Room{to_call: 40}

iex> room = %Room{to_call: 20}
iex> Updater.call_amount(room, 15)
%Room{to_call: 20}

iex> room = %Room{to_call: 20}
iex> Updater.call_amount(room, 20)
%Room{to_call: 20}

called(room, player)

called(PokerEx.Room2.t, player) :: PokerEx.Room2.t

Adds the given player to the front of the list of players who have called the current to_call amount.

Examples

iex> room = %Room{called: ["A", "B"]}
iex> Updater.called(room, "C")
%Room{called: ["C", "A", "B"]}

erase_called_with(room, player)

erase_called_with(PokerEx.Room2.t, player) :: PokerEx.Room2.t

Erases the list of players who have called and adds the specified player to the new list. This method is used when a player raises.

Examples

iex> room = %Room{called: ["A", "B", "C"]}
iex> Updater.erase_called_with(room, "D")
%Room{called: ["D"]}

insert_stats(room, player, number)

insert_stats(PokerEx.Room2.t, player, pos_integer) :: PokerEx.Room2.t

Allows you to insert an arbitrary score for a given player in the list of stats.

Examples

iex> Updater.insert_stats(%Room{}, "B", 100)
%Room{stats: [{"B", 100}]}

insert_winner(room, player)

insert_winner(PokerEx.Room2.t, player) :: PokerEx.Room2.t

Allows you to arbitrarily insert a winner.

Examples

iex> Updater.insert_winner(%Room{}, "A")
%Room{winner: "A"}

paid_in_round(room, player, amount)

paid_in_round(PokerEx.Room2.t, player, pos_integer) :: PokerEx.Room2.t

Updates the map stored under the round attribute on a room instance.

Examples

iex> room = %Room{round: %{"A" => 0}}
iex> Updater.paid_in_round(room, "A", 25)
%Room{round: %{"A" => 25}}

iex> room = %Room{round: %{"A" => 25}}
iex> Updater.paid_in_round(room, "A", 15)
%Room{round: %{"A" => 40}}

player_hands(room)

player_hands(PokerEx.Room2.t) :: PokerEx.Room2.t

Deals out hands and assigns each hand to active players.

Examples

iex> room = %Room{active: [{"A", 0}, {"B", 1}]}
iex> room = Updater.player_hands(room)
iex> length(room.player_hands) == 2
true

pot(room, amount)

pot(PokerEx.Room2.t, pos_integer) :: PokerEx.Room2.t

Updates the amount stored under the pot attribute on a room instance by adding the specified amount.

Examples

iex> room = %Room{pot: 75}
iex> Updater.pot(room, 25)
%Room{pot: 100}

iex> room = %Room{pot: 75}
iex> Updater.pot(room, -10)
%Room{pot: 75}

reindex_seating(room)

reindex_seating(PokerEx.Room2.t) :: PokerEx.Room2.t

Reindexes the indices on the seating attribute of a room instance to keep the seat_numbers consistent with each player’s position in the seating list

Examples

iex> room = %Room{seating: [{"A", 0}, {"C", 3}]}
iex> Updater.reindex_seating(room)
%Room{seating: [{"A", 0}, {"C", 1}]}

remove_from_seating(room, player)

remove_from_seating(PokerEx.Room2.t, player) :: PokerEx.Room2.t

Removes the specified player from the seating list.

Examples

iex> room = %Room{seating: [{"A", 0}, {"B", 1}, {"C", 2}]}
iex> Updater.remove_from_seating(room, "B")
%Room{seating: [{"A", 0}, {"C", 2}]}

remove_players_with_no_chips(room)

remove_players_with_no_chips(PokerEx.Room2.t) :: PokerEx.Room2.t

Removes players whose chip counts have fallen to zero and reindexes the list.

Examples

iex> PokerEx.AppState.put(PokerEx.Player.new("Michelangelo"))
iex> PokerEx.AppState.put(PokerEx.Player.new("Donatello"))
iex> PokerEx.Player.bet("Donatello", 1500)
iex> room = %Room{seating: [{"Donatello", 0}, {"Michelangelo", 1}]}
iex> room = Updater.remove_players_with_no_chips(room)
iex> room.seating
[{"Michelangelo", 0}]

reset_active(room)

reset_active(PokerEx.Room2.t) :: PokerEx.Room2.t

Resets the active list and sets the small_blind and big_blind as the last two players in the list, respectively.

Examples

iex> room = %Room{active: [{"A", 0}, {"B", 1}, {"C", 2}], current_big_blind: 1}
iex> room = Updater.reset_active(room)
iex> room.active
[{"C", 2}, {"A", 0}, {"B", 1}]

reset_all_in(room)

reset_all_in(PokerEx.Room2.t) :: PokerEx.Room2.t

Resets the list of all_in players to an empty list.

Examples

iex> room = %Room{all_in: ["A", "B", "C", "D"]}
iex> room = Updater.reset_all_in(room)
iex> room.all_in
[]

reset_call_amount(room)

reset_call_amount(PokerEx.Room2.t) :: PokerEx.Room2.t

Resets the to_call amount to zero.

Examples

iex> room = %Room{to_call: 500}
iex> room = Updater.reset_call_amount(room)
iex> room.to_call
0

reset_called(room)

reset_called(PokerEx.Room2.t) :: PokerEx.Room2.t

Resets the called list to an empty list

Examples

iex> room = %Room{called: ["A", "B", "C"]}
iex> room = Updater.reset_called(room)
iex> room.called
[]

reset_deck(room)

reset_deck(PokerEx.Room2.t) :: PokerEx.Room2.t

Resets the deck attribute to store a fresh, shuffled deck.

Examples

iex> deck = PokerEx.Deck.new |> PokerEx.Deck.shuffle
iex> room = %Room{deck: deck}
iex> room = Updater.reset_deck(room)
iex> deck == room.deck
false

reset_paid_in_round(room)

reset_paid_in_round(PokerEx.Room2.t) :: PokerEx.Room2.t

Resets the round map on a room instance to an empty map.

Examples

iex> room = %Room{round: %{"A" => 100, "B" => 50}}
iex> room = Updater.reset_paid_in_round(room)
iex> room.round
%{}

reset_player_hands(room)

reset_player_hands(PokerEx.Room2.t) :: PokerEx.Room2.t

Resets the player_hands attribute on a Room instance to an empty list.

Examples

iex> room = %Room{player_hands: {"A", [%PokerEx.Card{rank: :two, suit: :spades}, %PokerEx.Card{rank: :three, suit: :spades}]}}
iex> room = Updater.reset_player_hands(room)
iex> room.player_hands
[]

reset_pot(room)

reset_pot(PokerEx.Room2.t) :: PokerEx.Room2.t

Resets the pot attribute back to zero on a room instance.

Examples

iex> room = %Room{pot: 500}
iex> room = Updater.reset_pot(room)
iex> room.pot
0

reset_stats(room)

reset_stats(PokerEx.Room2.t) :: PokerEx.Room2.t

Resets the stats list on a room instance to an empty list.

Examples

iex> stats = [{"A", 400}, {"B", 600}]
iex> room = %Room{stats: stats}
iex> room = Updater.reset_stats(room)
iex> room.stats
[]

reset_table(room)

reset_table(PokerEx.Room2.t) :: PokerEx.Room2.t

Resets the table attribute on a room instance to an empty list.

Examples

iex> room = %Room{table: [%PokerEx.Card{rank: :two, suit: :spades}, %PokerEx.Card{rank: :three, suit: :spades}, 
...> %PokerEx.Card{rank: :four, suit: :spades}]}
iex> room = Updater.reset_table(room)
iex> room.table
[]

reset_total_paid(room)

reset_total_paid(PokerEx.Room2.t) :: PokerEx.Room2.t

Resets the :paid attribute on a room instance to an empty map.

Examples

iex> room = %Room{paid: %{"A" => 345, "B" => 700}}
iex> room = Updater.reset_total_paid(room)
iex> room.paid
%{}

reset_winner(room)

reset_winner(PokerEx.Room2.t) :: PokerEx.Room2.t

Resets the winner attribute on a room instance to nil

Examples

iex> room = %Room{winner: "A"}
iex> room = Updater.reset_winner(room)
iex> room.winner
nil

reset_winning_hand(room)

reset_winning_hand(PokerEx.Room2.t) :: PokerEx.Room2.t

Resets the winning_hand attribute on a room instance to nil.

Examples

iex> room = %Room{winning_hand: %PokerEx.Hand{score: 600, best_hand: []}}
iex> room = Updater.reset_winning_hand(room)
iex> room.winning_hand
nil

seating(room, player)

seating(PokerEx.Room2.t, player) :: PokerEx.Room2.t

Updates the seating attribute of a room instance

Examples

iex> room = %Room{seating: [{"A", 0}, {"B", 1}]}
iex> Updater.seating(room, "C")
%Room{seating: [{"A", 0}, {"B", 1}, {"C", 2}]}

set_active(room)

set_active(PokerEx.Room2.t) :: PokerEx.Room2.t

Sets the list of active players based on the seating list and the current big blind position

Examples

iex> room = %Room{seating: [{"A", 0}, {"B", 1}, {"C", 2}, {"D", 3}], current_big_blind: 2}
iex> room = Updater.set_active(room)
iex> room.active
[{"D", 3}, {"A", 0}, {"B", 1}, {"C", 2}]

stats(room)

stats(PokerEx.Room2.t) :: PokerEx.Room2.t

Updates the stats attribute on room instances given the player_hands and table cards.

Examples

iex> player_hands = [{"A", [%PokerEx.Card{rank: :two, suit: :hearts}, %PokerEx.Card{rank: :three, suit: :spades}]},
...> {"B", [%PokerEx.Card{rank: :four, suit: :clubs}, %PokerEx.Card{rank: :five, suit: :diamonds}]}]
iex> table = [%PokerEx.Card{rank: :ten, suit: :hearts}, %PokerEx.Card{rank: :jack, suit: :diamonds}, 
...> %PokerEx.Card{rank: :queen, suit: :spades}, %PokerEx.Card{rank: :king, suit: :clubs}, %PokerEx.Card{rank: :ace, suit: :spades}]
iex> room = %Room{player_hands: player_hands, table: table}
iex> result = Updater.stats(room)
iex> result.stats
[{"A", 414}, {"B", 414}]

table(room)

table(PokerEx.Room2.t) :: PokerEx.Room2.t

Deals a new card to be placed in the list of cards showing on the table.

Examples

iex> room = %Room{table: []}
iex> room = Updater.table(room)
iex> length(room.table) == 1
true

total_paid(room, player, amount)

total_paid(PokerEx.Room2.t, player, pos_integer) :: PokerEx.Room2.t

Updates the paid map stored under the paid attribute on a room instance for the specified player by adding the specified amount to the current amount.

Examples

iex> room = %Room{paid: %{"A" => 300}}
iex> Updater.total_paid(room, "A", 30)
%Room{paid: %{"A" => 330}}

winner(room)

winner(PokerEx.Room2.t) :: PokerEx.Room2.t

Evaluates player hands and determines the winner, then assigns the winner and winning_hand to their respective attributes in the room instance.

Examples

iex> player_hands = [{"A", [%PokerEx.Card{rank: :two, suit: :hearts}, %PokerEx.Card{rank: :three, suit: :hearts}]},
...> {"B", [%PokerEx.Card{rank: :four, suit: :clubs}, %PokerEx.Card{rank: :five, suit: :diamonds}]},
...> {"C", [%PokerEx.Card{rank: :six, suit: :diamonds}, %PokerEx.Card{rank: :seven, suit: :clubs}]}]
iex> table = [%PokerEx.Card{rank: :ten, suit: :hearts}, %PokerEx.Card{rank: :jack, suit: :diamonds}, 
...> %PokerEx.Card{rank: :queen, suit: :hearts}, %PokerEx.Card{rank: :king, suit: :hearts}, %PokerEx.Card{rank: :ace, suit: :spades}]
iex> active = [{"A", 0}]
iex> all_in = ["B"]
iex> room = %Room{player_hands: player_hands, table: table, active: active, all_in: all_in}
iex> room = Updater.stats(room)
iex> room = Updater.winner(room)
iex> {room.winner, room.winning_hand.type_string}
{"A", "a Flush, King High"}